Refactor client builder to not manipulate data
